pyproject.toml
[project]
name = "wfstats"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)"
]

[tool.poetry]
packages = [{include = "wfstats", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
